requirements.txt
pygame 
requests
pillow
ctype
pywin32
plyer
